+++ projects/net.splitcells.dem.merger/CHANGELOG.md
@@ -24,6 +24,7 @@
      This also makes it possible to use the name `Resource` for an interface
      of resource like things.
 ### Minor Changes
+* **2023-02-19** **\#170**: Log if most runtime performant settings are not active.
 * **2023-02-19** **\#10**: Create DescribedBool for assertions with descriptions.
 * **2023-02-02** **\#170**: Create container for arbitrary typed values with a type safe interface.
 * **2022-10-10** **\#8**: Create interface Identifiable, in order to be able to correctly implement `Object#equals` in wrappers and aspects.